requirements.txt
discord.py==2.6.4
python-dotenv==1.1.1
aiohttp==3.13.0
asyncio==4.0.0